🚚 api: second route: move session finalization usecase
to src directory
